+++ chk/test06.f90
@@ -544,7 +544,7 @@ CONTAINS
           DO ix = 1, 25, 8
             X = ix - 0.20D0
             CALL DEXINT(X,n,kode,m,tol,en,nz,ierr)
-            kx = X + 0.5D0
+            kx = INT( X + 0.5D0 )
             IF ( kx==0 ) kx = 1
             icase = 1
             A = n
@@ -637,9 +637,9 @@ CONTAINS
     xx(5) = 0.01D0
     DO i = 1, 5
       xx(i) = -xx(i)
-      k = xx(2)
-      n = xx(3)
-      m = xx(4)
+      k = INT( xx(2) )
+      n = INT( xx(3) )
+      m = INT( xx(4) )
       CALL DEXINT(xx(i),n,k,m,xx(5),en,nz,ierr)
       IF ( ierr/=1 ) THEN
         Ipass = 0
